% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTCSAM_AggCatchData.R
\name{extractTCSAM_AggCatchData}
\alias{extractTCSAM_AggCatchData}
\title{Extract aggregated abundance or biomass catch data for a TCSAM02 model run from a list of parsed text}
\usage{
extractTCSAM_AggCatchData(res, ks, verbose = FALSE)
}
\arguments{
\item{res}{- list of parsed text from a model fleet data file}

\item{ks}{- starting index into \code{res}}

\item{verbose}{- flag to print info}
}
\value{
list with aggregated abundance or biomass catch data
}
\description{
Function to extract aggregated abundance or biomass catch data for a TCSAM02 model run from a list of parsed text.
}
\details{
None.
}
