% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTCSAM_BoundedParameterInfo.R
\name{extractTCSAM_BoundedParameterInfo}
\alias{extractTCSAM_BoundedParameterInfo}
\title{Extract bounded parameters info from a parsed file for a TCSAM02 model run}
\usage{
extractTCSAM_BoundedParameterInfo(res, ks, param)
}
\arguments{
\item{res}{- list with parsed file results}

\item{ks}{- start location index}

\item{verbose}{- flag to print info}
}
\value{
list with elements
k - index of next element to parse
lst - list of model parameters info for recruitment parameters
}
\description{
Function to extract bounded parameters info from a parsed file for a TCSAM02 model run.
}
\details{
None.
}
