% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTCSAM_BoundedVectorInfo.R
\name{extractTCSAM_BoundedVectorInfo}
\alias{extractTCSAM_BoundedVectorInfo}
\title{Extract bounded vector parameters info from a parsed file for a TCSAM02 model run}
\usage{
extractTCSAM_BoundedVectorInfo(res, ks, param)
}
\arguments{
\item{res}{- list with parsed file results}

\item{ks}{- start location index}

\item{verbose}{- flag to print info}
}
\value{
list with elements
k - index of next element to parse
lst - list of model vector parameters info
}
\description{
Function to extract bounded vector parameters info from a parsed file for a TCSAM02 model run.
}
\details{
None.
}
