% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTCSAM_CatchData.R
\name{extractTCSAM_CatchData}
\alias{extractTCSAM_CatchData}
\title{Extract model catch data for a TCSAM02 model run from a list of parsed text}
\usage{
extractTCSAM_CatchData(res, ks, verbose = FALSE)
}
\arguments{
\item{res}{- list of parsed text from model fleet data file}

\item{ks}{- starting index into \code{res}}

\item{verbose}{- flag to print info}
}
\value{
list with catch data
}
\description{
Function to extract model catch data for a TCSAM02 model run from a list of parsed text.
}
\details{
None.
}
