% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTCSAM_EffortData.R
\name{extractTCSAM_EffortData}
\alias{extractTCSAM_EffortData}
\title{Extract catch effort data for a TCSAM02 model run from a list of parsed text}
\usage{
extractTCSAM_EffortData(res, ks, verbose = FALSE)
}
\arguments{
\item{res}{- list of parsed text from a model fleet data file}

\item{ks}{- starting index into \code{res}}

\item{verbose}{- flag to print info}
}
\value{
list with catch effort data
}
\description{
Function to extract catch effort data for a TCSAM02 model run from a list of parsed text.
}
\details{
None.
}
