% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTCSAM_ParamInfoNaturalMortality.R
\name{extractTCSAM_ParamInfoNaturalMortality}
\alias{extractTCSAM_ParamInfoNaturalMortality}
\title{Extract model natural mortality parameters info from a parsed file for a TCSAM02 model run}
\usage{
extractTCSAM_ParamInfoNaturalMortality(res, ks, verbose = FALSE)
}
\arguments{
\item{res}{- list with parsed file results}

\item{ks}{- start location index}

\item{verbose}{- flag to print info}
}
\value{
list with elements
k - index of next element to parse
lst - list of model parameters info for natural mortality parameters
}
\description{
Function to extract model natural mortality parameters info from a parsed file for a TCSAM02 model run.
}
\details{
None.
}
