% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTCSAM_PriorInfo.R
\name{extractTCSAM_PriorInfo}
\alias{extractTCSAM_PriorInfo}
\title{Extract prior probability info from a character vector for a TCSAM02 model run}
\usage{
extractTCSAM_PriorInfo(pv, k, verbose = FALSE)
}
\arguments{
\item{pv}{- character vector with parameter vector information}

\item{k}{- start location index for prior information}

\item{verbose}{- flag to print info}
}
\value{
list with elements
wgt - prior weight
type - distribution type for prior
params - numeric vector with parameters for prior
consts - numeric vector with constants for prior
}
\description{
Function to extract prior probability info from a character vector for a TCSAM02 model run.
}
\details{
None.
}
