% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTCSAM_ZCsCatchData.R
\name{extractTCSAM_ZCsCatchData}
\alias{extractTCSAM_ZCsCatchData}
\title{Extract size composition data for a TCSAM02 model run from a list of parsed text}
\usage{
extractTCSAM_ZCsCatchData(res, ks, verbose = FALSE)
}
\arguments{
\item{res}{- list of parsed text from a model fleet data file}

\item{ks}{- starting index into \code{res}}

\item{verbose}{- flag to print info}
}
\value{
list with size composition data
}
\description{
Function to extract size composition data for a TCSAM02 model run from a list of parsed text.
}
\details{
None.
}
