% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseTimeBlock.R
\name{parseTimeBlock}
\alias{parseTimeBlock}
\title{Parse a formatted time block}
\usage{
parseTimeBlock(str, verbose = FALSE)
}
\arguments{
\item{str}{- the time block (as text) to parse}

\item{verbose}{- flag to print info}
}
\value{
numeric vector with parsed time block
}
\description{
Function to parse a formatted time block.
}
\details{
The input time block should follow one of the following formats:
\code{"1.23"}
\code{"100:110"}
\code{"[110:120]"}
\code{"[110:120; 130]"}
\code{"[110:120; 130:140]"}
}
