% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTCSAM_Datasets.R
\name{readTCSAM_Datasets}
\alias{readTCSAM_Datasets}
\title{Read model datasets file for a TCSAM02 model run}
\usage{
readTCSAM_Datasets(conn)
}
\arguments{
\item{fn}{- file name for model datasets file}

\item{verbose}{- flag to print info}
}
\value{
list with input data sets.
}
\description{
Function to read model datasets file for a TCSAM02 model run.
}
\details{
None.
}
