% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTCSAM_FleetData.R
\name{readTCSAM_FleetData}
\alias{readTCSAM_FleetData}
\title{Read model fleet data file for a TCSAM02 model run}
\usage{
readTCSAM_FleetData(conn, verbose = FALSE)
}
\arguments{
\item{conn}{- file name for model fleet data file}

\item{verbose}{- flag to print info}
}
\value{
list with fleet data set
}
\description{
Function to read model fleet data file for a TCSAM02 model run.
}
\details{
None.
}
