% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTCSAM_GrowthData.R
\name{readTCSAM_GrowthData}
\alias{readTCSAM_GrowthData}
\title{Read model growth data file for a TCSAM02 model run}
\usage{
readTCSAM_GrowthData(conn)
}
\arguments{
\item{fn}{- file name for model growth data file}

\item{verbose}{- flag to print info}
}
\value{
list with growth data set
}
\description{
Function to read model growth data file for a TCSAM02 model run.
}
\details{
None.
}
