% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTCSAM.R
\name{readTCSAM_InputFiles}
\alias{readTCSAM_InputFiles}
\title{Read input files for a TCSAM02 model run}
\usage{
readTCSAM_InputFiles(fn, verbose = TRUE)
}
\arguments{
\item{fn}{- file name for model configuration file}

\item{verbose}{- flag to print info}
}
\value{
list with input files information
}
\description{
Function to read input files for a TCSAM02 model run.
}
\details{
None.
}
