% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTCSAM_MaturityOgiveData.R
\name{readTCSAM_MaturityOgiveData}
\alias{readTCSAM_MaturityOgiveData}
\title{Read model maturity ogive data file for a TCSAM02 model run}
\usage{
readTCSAM_MaturityOgiveData(conn)
}
\arguments{
\item{fn}{- file name for model maturity ogive data file}

\item{verbose}{- flag to print info}
}
\value{
list with maturity ogive data set
}
\description{
Function to read model maturity ogive data file for a TCSAM02 model run.
}
\details{
None.
}
