% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTCSAM_ModelConfiguration.R
\name{readTCSAM_ModelConfiguration}
\alias{readTCSAM_ModelConfiguration}
\title{Read model configuration file for a TCSAM02 model run}
\usage{
readTCSAM_ModelConfiguration(conn, verbose = FALSE)
}
\arguments{
\item{verbose}{- flag to print info}

\item{fn}{- file name for model configuration file}
}
\value{
list with model configuration information
}
\description{
Function to read model configuration file for a TCSAM02 model run.
}
\details{
None.
}
