% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTCSAM_ModelParametersInfo.R
\name{readTCSAM_ModelParametersInfo}
\alias{readTCSAM_ModelParametersInfo}
\title{Read model parameters info file for a TCSAM02 model run}
\usage{
readTCSAM_ModelParametersInfo(conn, verbose = FALSE)
}
\arguments{
\item{conn}{- file name for model parameters info file}

\item{verbose}{- flag to print info}
}
\value{
list with model parameters info
}
\description{
Function to read model parameters info file for a TCSAM02 model run.
}
\details{
None.
}
