% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGlobals.R
\name{setGlobals}
\alias{setGlobals}
\title{Set global values}
\usage{
setGlobals()
}
\description{
Function to set global values, like key/value maps and dimensions.
}
\details{
Must be called prior to any other functions. The following global values are created:
\itemize{
  \item{.KV_MAP - list of named value vectors with keys as the associated names}
  \item{.DIMS - list of model dimensions}
  \item{.DIMNAMES - list of names associated with the model dimensions}
}
}
