% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGlobals.R
\name{subForKey}
\alias{subForKey}
\title{Substitute value for key}
\usage{
subForKey(ks, map_kv)
}
\arguments{
\item{ks}{- vector of key words}

\item{map_kv}{- map from keyword to value}
}
\description{
Function to substitute values for key words.
}
\details{
None.
}
